requirement.txt
Django==1.9
django-bootstrap3
django-ipware
djangorestframework
django-cors-headers
drf-nested-routers
